feat: add getDataOrders function modularization
